requirements.txt
discord
bs4
pytube3
PyNaCl
ffmpeg